+++ src/view.c
@@ -136,41 +136,46 @@ static inline float window_node_get_gap(struct view *view)
 }
 
 #define area_ax_truncate(a)  \
-    a.x = (int)(a.x + 0.5f); \
-    a.y = (int)(a.y + 0.5f); \
-    a.w = (int)(a.w + 0.5f); \
-    a.h = (int)(a.h + 0.5f);
+    a->x = (int)(a->x + 0.5f); \
+    a->y = (int)(a->y + 0.5f); \
+    a->w = (int)(a->w + 0.5f); \
+    a->h = (int)(a->h + 0.5f);
 
-static void area_make_pair(struct view *view, struct window_node *node)
+static void area_make_pair(enum window_node_split split, float gap, float ratio, struct area *parent_area, struct area *left_area, struct area *right_area)
+{
+    if (split == SPLIT_Y) {
+        *left_area = *parent_area;
+        left_area->w *= ratio;
+        left_area->w -= gap;
+
+        *right_area = *parent_area;
+        right_area->x += (parent_area->w * ratio);
+        right_area->w *= (1 - ratio);
+        right_area->x += gap;
+        right_area->w -= gap;
+    } else {
+        *left_area = *parent_area;
+        left_area->h *= ratio;
+        left_area->h -= gap;
+
+        *right_area = *parent_area;
+        right_area->y += (parent_area->h * ratio);
+        right_area->h *= (1 - ratio);
+        right_area->y += gap;
+        right_area->h -= gap;
+    }
+
+    area_ax_truncate(left_area);
+    area_ax_truncate(right_area);
+}
+
+static void area_make_pair_for_node(struct view *view, struct window_node *node)
 {
     enum window_node_split split = window_node_get_split(node);
     float ratio = window_node_get_ratio(node);
     float gap   = window_node_get_gap(view);
 
-    if (split == SPLIT_Y) {
-        node->left->area = node->area;
-        node->left->area.w *= ratio;
-        node->left->area.w -= gap;
-
-        node->right->area = node->area;
-        node->right->area.x += (node->area.w * ratio);
-        node->right->area.w *= (1 - ratio);
-        node->right->area.x += gap;
-        node->right->area.w -= gap;
-    } else {
-        node->left->area = node->area;
-        node->left->area.h *= ratio;
-        node->left->area.h -= gap;
-
-        node->right->area = node->area;
-        node->right->area.y += (node->area.h * ratio);
-        node->right->area.h *= (1 - ratio);
-        node->right->area.y += gap;
-        node->right->area.h -= gap;
-    }
-
-    area_ax_truncate(node->left->area);
-    area_ax_truncate(node->right->area);
+    area_make_pair(split, gap, ratio, &node->area, &node->left->area, &node->right->area);
 
     node->split = split;
     node->ratio = ratio;
@@ -283,13 +288,13 @@ static void window_node_split(struct view *view, struct window_node *node, struc
     node->right = right;
     node->zoom  = NULL;
 
-    area_make_pair(view, node);
+    area_make_pair_for_node(view, node);
 }
 
 void window_node_update(struct view *view, struct window_node *node)
 {
     if (window_node_is_intermediate(node)) {
-        area_make_pair(view, node->parent);
+        area_make_pair_for_node(view, node->parent);
     }
 
     if (window_node_is_leaf(node)) {